+++ contracts/ethereum/testContracts/NamedStorage.sol
@@ -1,0 +1,93 @@
+// SPDX-License-Identifier: Apache-2.0.
+pragma solidity ^0.8.6;
+
+/*
+  Library to provide basic storage, in storage location out of the low linear address space.
+  New types of storage variables should be added here upon need.
+*/
+library NamedStorage {
+  function bytes32ToUint256Mapping(string memory tag_)
+    internal
+    pure
+    returns (mapping(bytes32 => uint256) storage randomVariable)
+  {
+    bytes32 location = keccak256(abi.encodePacked(tag_));
+    assembly {
+      randomVariable.slot := location
+    }
+  }
+
+  function bytes32ToAddressMapping(string memory tag_)
+    internal
+    pure
+    returns (mapping(bytes32 => address) storage randomVariable)
+  {
+    bytes32 location = keccak256(abi.encodePacked(tag_));
+    assembly {
+      randomVariable.slot := location
+    }
+  }
+
+  function addressToBoolMapping(string memory tag_)
+    internal
+    pure
+    returns (mapping(address => bool) storage randomVariable)
+  {
+    bytes32 location = keccak256(abi.encodePacked(tag_));
+    assembly {
+      randomVariable.slot := location
+    }
+  }
+
+  function getUintValue(string memory tag_) internal view returns (uint256 retVal) {
+    bytes32 slot = keccak256(abi.encodePacked(tag_));
+    assembly {
+      retVal := sload(slot)
+    }
+  }
+
+  function setUintValue(string memory tag_, uint256 value) internal {
+    bytes32 slot = keccak256(abi.encodePacked(tag_));
+    assembly {
+      sstore(slot, value)
+    }
+  }
+
+  function setUintValueOnce(string memory tag_, uint256 value) internal {
+    require(getUintValue(tag_) == 0, 'ALREADY_SET');
+    setUintValue(tag_, value);
+  }
+
+  function getAddressValue(string memory tag_) internal view returns (address retVal) {
+    bytes32 slot = keccak256(abi.encodePacked(tag_));
+    assembly {
+      retVal := sload(slot)
+    }
+  }
+
+  function setAddressValue(string memory tag_, address value) internal {
+    bytes32 slot = keccak256(abi.encodePacked(tag_));
+    assembly {
+      sstore(slot, value)
+    }
+  }
+
+  function setAddressValueOnce(string memory tag_, address value) internal {
+    require(getAddressValue(tag_) == address(0x0), 'ALREADY_SET');
+    setAddressValue(tag_, value);
+  }
+
+  function getBoolValue(string memory tag_) internal view returns (bool retVal) {
+    bytes32 slot = keccak256(abi.encodePacked(tag_));
+    assembly {
+      retVal := sload(slot)
+    }
+  }
+
+  function setBoolValue(string memory tag_, bool value) internal {
+    bytes32 slot = keccak256(abi.encodePacked(tag_));
+    assembly {
+      sstore(slot, value)
+    }
+  }
+}
