+++ contracts/starknet/account/ERC165_base.cairo
@@ -10,7 +10,8 @@ func ERC165_supported_interfaces(interface_id : felt) -> (is_supported : felt):
 end
 
 func ERC165_supports_interface{syscall_ptr : felt*, pedersen_ptr : HashBuiltin*, range_check_ptr}(
-        interface_id : felt) -> (success : felt):
+    interface_id : felt
+) -> (success : felt):
     # 165
     if interface_id == 0x01ffc9a7:
         return (TRUE)
@@ -22,7 +23,8 @@ func ERC165_supports_interface{syscall_ptr : felt*, pedersen_ptr : HashBuiltin*,
 end
 
 func ERC165_register_interface{syscall_ptr : felt*, pedersen_ptr : HashBuiltin*, range_check_ptr}(
-        interface_id : felt):
+    interface_id : felt
+):
     # Ensures interface_id is not the invalid interface_id
     assert_not_equal(interface_id, 0xffffffff)
     ERC165_supported_interfaces.write(interface_id, TRUE)
