+++ contracts/starknet/authenticator/vanilla.cairo
@@ -4,7 +4,8 @@ from starkware.starknet.common.syscalls import call_contract
 # Forwards `data` to `target` without verifying anything.
 @external
 func execute{syscall_ptr : felt*, range_check_ptr}(
-        to : felt, function_selector : felt, calldata_len : felt, calldata : felt*) -> ():
+    to : felt, function_selector : felt, calldata_len : felt, calldata : felt*
+) -> ():
     # TODO: Actually verify the signature
 
     # Call the contract
@@ -12,7 +13,8 @@ func execute{syscall_ptr : felt*, range_check_ptr}(
         contract_address=to,
         function_selector=function_selector,
         calldata_size=calldata_len,
-        calldata=calldata)
+        calldata=calldata,
+    )
 
     return ()
 end
