+++ contracts/starknet/testContracts/test_words.cairo
@@ -4,7 +4,8 @@ from starknet.lib.words import Words, felt_to_words
 
 @view
 func test_felt_to_words{range_check_ptr, bitwise_ptr : BitwiseBuiltin*}(input : felt) -> (
-        words : Words):
+    words : Words
+):
     let (words) = felt_to_words(input)
     return (words)
 end
