+++ package.json
@@ -8,16 +8,10 @@
     "compile:l2": "hardhat starknet-compile",
     "compile": "yarn compile:l1 && yarn compile:l2",
     "deploy": "sh deploy.sh",
-<<<<<<< HEAD
-    "test:hardhat": "yarn hardhat test test/ethereum/*.ts test/starknet/*.ts --network 'hardhat' --starknet-network 'starknetLocal'",
-    "test:l1l2": "yarn hardhat test test/l1l2/*.ts --network 'ethereumLocal' --starknet-network 'starknetLocal'",
-    "test": "yarn test:hardhat && yarn test:l1l2",
-=======
     "test:l1": "yarn hardhat test test/ethereum/*.ts --network 'hardhat' --starknet-network 'starknetLocal'",
     "test:l2": "yarn hardhat test test/starknet/*.ts --network 'hardhat' --starknet-network 'starknetLocal'",
     "test:crosschain": "yarn hardhat test test/crosschain/*.ts --network 'ethereumLocal' --starknet-network 'starknetLocal'",
     "test": "yarn test:l1 && yarn test:l2 && yarn test:crosschain",
->>>>>>> develop
     "start:l1": "hardhat node",
     "format:l1": "prettier --write 'contracts/**/*.sol'",
     "check-format:l1": "prettier -c 'contracts/**/*.sol'",
